// src/interfaces/global.ts
type Applicant = {
  id: string;
  userId: string;
  firstName: string;
  lastName: string;
  email: string;
  profilePicture: string | null;
  companyInfo: any | null; // You might want to replace 'any' with a more specific type
  createdAt?: string;
  updatedAt?: string;
};

export type JobApplication = {
  id: string;
  userId: string;
  jobId: string;
  createdAt: string;
  updatedAt: string;
  user: Applicant;
};

// Types for Job data
export type Job = {
  id: string;
  userId: string;
  jobPosition: string;
  yourPosition: string;
  location: string;
  jobType: string;
  salaryRange: string;
  experienceLevel: string;
  aboutTheJob: string;
  responsibilities: string;
  reqQualifications: string;
  reqSkills: string[];
  companyName: string | null;
  aboutTheCompany: string | null;
  companyLogo: string | null;
  workScheduleBenefits: string | null;
  createdAt: string;
  updatedAt: string;
  totalApplicants?: number;
  isPaid: boolean;
  ApplyJob?: JobApplication[];
  user?: {
    id: string;
    firstName: string;
    lastName: string;
    email: string;
    profilePicture: string | null;
    companyInfo?: {
      id: string;
      userId: string;
      companyName: string;
      companyDetails: string;
      phoneNumber: string;
      address: string;
      establish: string | null;
      createdAt: string;
      updatedAt: string;
    };
  };
  isFavorite: boolean;
};